+++ src/frontend/apps/drive/src/features/explorer/components/trash/utils.tsx
@@ -2,13 +2,15 @@ import { useModals } from "@gouvfr-lasuite/cunningham-react";
 import i18n from "@/features/i18n/initI18n";
 
 export const messageModalTrashNavigate = (
-  modals: ReturnType<typeof useModals>
+  modals: ReturnType<typeof useModals>,
+  isFile: boolean = false,
 ) => {
+  const key = isFile ? "modal_file" : "modal_folder";
   modals.messageModal({
-    title: i18n.t("explorer.trash.navigate.modal.title"),
+    title: i18n.t(`explorer.trash.navigate.${key}.title`),
     children: (
       <div className="clr-greyscale-600">
-        {i18n.t("explorer.trash.navigate.modal.description")}
+        {i18n.t(`explorer.trash.navigate.${key}.description`)}
       </div>
     ),
   });